+++ public/assets/js/getUpload.js
@@ -1,4 +1,4 @@
-import { createFormAndUpdateImage } from "./createFormAndUpdateImage.js";
+import { showUpdateImage } from "./showUpdateImage.js";
 import { deleteImg } from "./deleteImg.js";
 import { showImageUpload } from "./showImageUpload.js";
 
@@ -7,13 +7,20 @@ const urlGetImages = "/images";
 
 fetch(`${apiBaseUrl}${urlGetImages}`)
     .then((response) => response.json())
-    .catch(error => console.error('Error:', error))
+    .catch(error => {
+        console.error('Error:', error);
+        Swal.fire({
+            icon: 'error',
+            title: 'Oops...algo salió mal',
+            text: `${error}`,
+        })
+    })
     .then((contenido) => {
         let images = contenido.image;
 
         showImageUpload(images);
 
-        createFormAndUpdateImage(images);
+        showUpdateImage(images);
 
         deleteImg(images);
